65C02 Opcode upgrade
