+++ cpu/instructions.h
@@ -1,0 +1,490 @@
+/*
+
+						Extracted from original single fake6502.c file
+
+*/
+//
+//          65C02 changes.
+//
+//          BRK                 now clears D
+//          ADC/SBC             set N and Z in decimal mode. They also set V, but this is
+//                              essentially meaningless so this has not been implemented.
+//
+//
+//
+//          instruction handler functions
+//
+static void adc() {
+    penaltyop = 1;
+    value = getvalue();
+    result = (uint16_t)a + value + (uint16_t)(status & FLAG_CARRY);
+   
+    carrycalc(result);
+    zerocalc(result);
+    overflowcalc(result, a, value);
+    signcalc(result);
+    
+    #ifndef NES_CPU
+    if (status & FLAG_DECIMAL) {
+        clearcarry();
+        
+        if ((a & 0x0F) > 0x09) {
+            a += 0x06;
+        }
+        if ((a & 0xF0) > 0x90) {
+            a += 0x60;
+            setcarry();
+        }
+
+        zerocalc(a);                /* 65C02 change, Decimal Arithmetic sets NZV */
+        signcalc(a);
+        
+        clockticks6502++;
+    }
+    #endif
+   
+    saveaccum(result);
+}
+
+static void and() {
+    penaltyop = 1;
+    value = getvalue();
+    result = (uint16_t)a & value;
+   
+    zerocalc(result);
+    signcalc(result);
+   
+    saveaccum(result);
+}
+
+static void asl() {
+    value = getvalue();
+    result = value << 1;
+
+    carrycalc(result);
+    zerocalc(result);
+    signcalc(result);
+   
+    putvalue(result);
+}
+
+static void bcc() {
+    if ((status & FLAG_CARRY) == 0) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void bcs() {
+    if ((status & FLAG_CARRY) == FLAG_CARRY) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void beq() {
+    if ((status & FLAG_ZERO) == FLAG_ZERO) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void bit() {
+    value = getvalue();
+    result = (uint16_t)a & value;
+   
+    zerocalc(result);
+    status = (status & 0x3F) | (uint8_t)(value & 0xC0);
+}
+
+static void bmi() {
+    if ((status & FLAG_SIGN) == FLAG_SIGN) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void bne() {
+    if ((status & FLAG_ZERO) == 0) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void bpl() {
+    if ((status & FLAG_SIGN) == 0) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void brk() {
+    pc++;
+
+        
+    push16(pc); //push next instruction address onto stack
+    push8(status | FLAG_BREAK); //push CPU status to stack
+    setinterrupt(); //set interrupt flag
+    cleardecimal();       // clear decimal flag (65C02 change)
+    pc = (uint16_t)read6502(0xFFFE) | ((uint16_t)read6502(0xFFFF) << 8);
+}
+
+static void bvc() {
+    if ((status & FLAG_OVERFLOW) == 0) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void bvs() {
+    if ((status & FLAG_OVERFLOW) == FLAG_OVERFLOW) {
+        oldpc = pc;
+        pc += reladdr;
+        if ((oldpc & 0xFF00) != (pc & 0xFF00)) clockticks6502 += 2; //check if jump crossed a page boundary
+            else clockticks6502++;
+    }
+}
+
+static void clc() {
+    clearcarry();
+}
+
+static void cld() {
+    cleardecimal();
+}
+
+static void cli() {
+    clearinterrupt();
+}
+
+static void clv() {
+    clearoverflow();
+}
+
+static void cmp() {
+    penaltyop = 1;
+    value = getvalue();
+    result = (uint16_t)a - value;
+   
+    if (a >= (uint8_t)(value & 0x00FF)) setcarry();
+        else clearcarry();
+    if (a == (uint8_t)(value & 0x00FF)) setzero();
+        else clearzero();
+    signcalc(result);
+}
+
+static void cpx() {
+    value = getvalue();
+    result = (uint16_t)x - value;
+   
+    if (x >= (uint8_t)(value & 0x00FF)) setcarry();
+        else clearcarry();
+    if (x == (uint8_t)(value & 0x00FF)) setzero();
+        else clearzero();
+    signcalc(result);
+}
+
+static void cpy() {
+    value = getvalue();
+    result = (uint16_t)y - value;
+   
+    if (y >= (uint8_t)(value & 0x00FF)) setcarry();
+        else clearcarry();
+    if (y == (uint8_t)(value & 0x00FF)) setzero();
+        else clearzero();
+    signcalc(result);
+}
+
+static void dec() {
+    value = getvalue();
+    result = value - 1;
+   
+    zerocalc(result);
+    signcalc(result);
+   
+    putvalue(result);
+}
+
+static void dex() {
+    x--;
+   
+    zerocalc(x);
+    signcalc(x);
+}
+
+static void dey() {
+    y--;
+   
+    zerocalc(y);
+    signcalc(y);
+}
+
+static void eor() {
+    penaltyop = 1;
+    value = getvalue();
+    result = (uint16_t)a ^ value;
+   
+    zerocalc(result);
+    signcalc(result);
+   
+    saveaccum(result);
+}
+
+static void inc() {
+    value = getvalue();
+    result = value + 1;
+   
+    zerocalc(result);
+    signcalc(result);
+   
+    putvalue(result);
+}
+
+static void inx() {
+    x++;
+   
+    zerocalc(x);
+    signcalc(x);
+}
+
+static void iny() {
+    y++;
+   
+    zerocalc(y);
+    signcalc(y);
+}
+
+static void jmp() {
+    pc = ea;
+}
+
+static void jsr() {
+    push16(pc - 1);
+    pc = ea;
+}
+
+static void lda() {
+    penaltyop = 1;
+    value = getvalue();
+    a = (uint8_t)(value & 0x00FF);
+   
+    zerocalc(a);
+    signcalc(a);
+}
+
+static void ldx() {
+    penaltyop = 1;
+    value = getvalue();
+    x = (uint8_t)(value & 0x00FF);
+   
+    zerocalc(x);
+    signcalc(x);
+}
+
+static void ldy() {
+    penaltyop = 1;
+    value = getvalue();
+    y = (uint8_t)(value & 0x00FF);
+   
+    zerocalc(y);
+    signcalc(y);
+}
+
+static void lsr() {
+    value = getvalue();
+    result = value >> 1;
+   
+    if (value & 1) setcarry();
+        else clearcarry();
+    zerocalc(result);
+    signcalc(result);
+   
+    putvalue(result);
+}
+
+static void nop() {
+    switch (opcode) {
+        case 0x1C:
+        case 0x3C:
+        case 0x5C:
+        case 0x7C:
+        case 0xDC:
+        case 0xFC:
+            penaltyop = 1;
+            break;
+    }
+}
+
+static void ora() {
+    penaltyop = 1;
+    value = getvalue();
+    result = (uint16_t)a | value;
+   
+    zerocalc(result);
+    signcalc(result);
+   
+    saveaccum(result);
+}
+
+static void pha() {
+    push8(a);
+}
+
+static void php() {
+    push8(status | FLAG_BREAK);
+}
+
+static void pla() {
+    a = pull8();
+   
+    zerocalc(a);
+    signcalc(a);
+}
+
+static void plp() {
+    status = pull8() | FLAG_CONSTANT;
+}
+
+static void rol() {
+    value = getvalue();
+    result = (value << 1) | (status & FLAG_CARRY);
+   
+    carrycalc(result);
+    zerocalc(result);
+    signcalc(result);
+   
+    putvalue(result);
+}
+
+static void ror() {
+    value = getvalue();
+    result = (value >> 1) | ((status & FLAG_CARRY) << 7);
+   
+    if (value & 1) setcarry();
+        else clearcarry();
+    zerocalc(result);
+    signcalc(result);
+   
+    putvalue(result);
+}
+
+static void rti() {
+    status = pull8();
+    value = pull16();
+    pc = value;
+}
+
+static void rts() {
+    value = pull16();
+    pc = value + 1;
+}
+
+static void sbc() {
+    penaltyop = 1;
+    value = getvalue() ^ 0x00FF;
+    result = (uint16_t)a + value + (uint16_t)(status & FLAG_CARRY);
+   
+    carrycalc(result);
+    zerocalc(result);
+    overflowcalc(result, a, value);
+    signcalc(result);
+
+    #ifndef NES_CPU
+    if (status & FLAG_DECIMAL) {
+        clearcarry();
+        
+        a -= 0x66;
+        if ((a & 0x0F) > 0x09) {
+            a += 0x06;
+        }
+        if ((a & 0xF0) > 0x90) {
+            a += 0x60;
+            setcarry();
+        }
+        
+        zerocalc(a);                /* 65C02 change, Decimal Arithmetic sets NZV */
+        signcalc(a);
+
+        clockticks6502++;
+    }
+    #endif
+   
+    saveaccum(result);
+}
+
+static void sec() {
+    setcarry();
+}
+
+static void sed() {
+    setdecimal();
+}
+
+static void sei() {
+    setinterrupt();
+}
+
+static void sta() {
+    putvalue(a);
+}
+
+static void stx() {
+    putvalue(x);
+}
+
+static void sty() {
+    putvalue(y);
+}
+
+static void tax() {
+    x = a;
+   
+    zerocalc(x);
+    signcalc(x);
+}
+
+static void tay() {
+    y = a;
+   
+    zerocalc(y);
+    signcalc(y);
+}
+
+static void tsx() {
+    x = sp;
+   
+    zerocalc(x);
+    signcalc(x);
+}
+
+static void txa() {
+    a = x;
+   
+    zerocalc(a);
+    signcalc(a);
+}
+
+static void txs() {
+    sp = x;
+}
+
+static void tya() {
+    a = y;
+   
+    zerocalc(a);
+    signcalc(a);
+}
