+++ cpu/modes.h
@@ -1,0 +1,102 @@
+/*
+
+						Extracted from original single fake6502.c file
+
+*/
+
+//
+//          65C02 changes.
+//
+//          ind         absolute indirect
+//
+//                      A 6502 has a bug whereby if you jmp ($12FF) it reads the address from
+//                      $12FF and $1200. This has been fixed in the 65C02. 
+//                      
+static void imp() { //implied
+}
+
+static void acc() { //accumulator
+}
+
+static void imm() { //immediate
+    ea = pc++;
+}
+
+static void zp() { //zero-page
+    ea = (uint16_t)read6502((uint16_t)pc++);
+}
+
+static void zpx() { //zero-page,X
+    ea = ((uint16_t)read6502((uint16_t)pc++) + (uint16_t)x) & 0xFF; //zero-page wraparound
+}
+
+static void zpy() { //zero-page,Y
+    ea = ((uint16_t)read6502((uint16_t)pc++) + (uint16_t)y) & 0xFF; //zero-page wraparound
+}
+
+static void rel() { //relative for branch ops (8-bit immediate value, sign-extended)
+    reladdr = (uint16_t)read6502(pc++);
+    if (reladdr & 0x80) reladdr |= 0xFF00;
+}
+
+static void abso() { //absolute
+    ea = (uint16_t)read6502(pc) | ((uint16_t)read6502(pc+1) << 8);
+    pc += 2;
+}
+
+static void absx() { //absolute,X
+    uint16_t startpage;
+    ea = ((uint16_t)read6502(pc) | ((uint16_t)read6502(pc+1) << 8));
+    startpage = ea & 0xFF00;
+    ea += (uint16_t)x;
+
+    if (startpage != (ea & 0xFF00)) { //one cycle penlty for page-crossing on some opcodes
+        penaltyaddr = 1;
+    }
+
+    pc += 2;
+}
+
+static void absy() { //absolute,Y
+    uint16_t startpage;
+    ea = ((uint16_t)read6502(pc) | ((uint16_t)read6502(pc+1) << 8));
+    startpage = ea & 0xFF00;
+    ea += (uint16_t)y;
+
+    if (startpage != (ea & 0xFF00)) { //one cycle penlty for page-crossing on some opcodes
+        penaltyaddr = 1;
+    }
+
+    pc += 2;
+}
+
+static void ind() { //indirect
+    uint16_t eahelp, eahelp2;
+    eahelp = (uint16_t)read6502(pc) | (uint16_t)((uint16_t)read6502(pc+1) << 8);
+    //
+    //      The 6502 page boundary wraparound bug does not occur on a 65C02.
+    //
+    //eahelp2 = (eahelp & 0xFF00) | ((eahelp + 1) & 0x00FF); //replicate 6502 page-boundary wraparound bug
+    eahelp2 = (eahelp+1) & 0xFFFF;
+    ea = (uint16_t)read6502(eahelp) | ((uint16_t)read6502(eahelp2) << 8);
+    pc += 2;
+}
+
+static void indx() { // (indirect,X)
+    uint16_t eahelp;
+    eahelp = (uint16_t)(((uint16_t)read6502(pc++) + (uint16_t)x) & 0xFF); //zero-page wraparound for table pointer
+    ea = (uint16_t)read6502(eahelp & 0x00FF) | ((uint16_t)read6502((eahelp+1) & 0x00FF) << 8);
+}
+
+static void indy() { // (indirect),Y
+    uint16_t eahelp, eahelp2, startpage;
+    eahelp = (uint16_t)read6502(pc++);
+    eahelp2 = (eahelp & 0xFF00) | ((eahelp + 1) & 0x00FF); //zero-page wraparound
+    ea = (uint16_t)read6502(eahelp) | ((uint16_t)read6502(eahelp2) << 8);
+    startpage = ea & 0xFF00;
+    ea += (uint16_t)y;
+
+    if (startpage != (ea & 0xFF00)) { //one cycle penlty for page-crossing on some opcodes
+        penaltyaddr = 1;
+    }
+}
