+++ cpu/support.h
@@ -1,0 +1,76 @@
+/*
+
+						Extracted from original single fake6502.c file
+
+*/
+
+#define saveaccum(n) a = (uint8_t)((n) & 0x00FF)
+
+
+//flag modifier macros
+#define setcarry() status |= FLAG_CARRY
+#define clearcarry() status &= (~FLAG_CARRY)
+#define setzero() status |= FLAG_ZERO
+#define clearzero() status &= (~FLAG_ZERO)
+#define setinterrupt() status |= FLAG_INTERRUPT
+#define clearinterrupt() status &= (~FLAG_INTERRUPT)
+#define setdecimal() status |= FLAG_DECIMAL
+#define cleardecimal() status &= (~FLAG_DECIMAL)
+#define setoverflow() status |= FLAG_OVERFLOW
+#define clearoverflow() status &= (~FLAG_OVERFLOW)
+#define setsign() status |= FLAG_SIGN
+#define clearsign() status &= (~FLAG_SIGN)
+
+
+//flag calculation macros
+#define zerocalc(n) {\
+    if ((n) & 0x00FF) clearzero();\
+        else setzero();\
+}
+
+#define signcalc(n) {\
+    if ((n) & 0x0080) setsign();\
+        else clearsign();\
+}
+
+#define carrycalc(n) {\
+    if ((n) & 0xFF00) setcarry();\
+        else clearcarry();\
+}
+
+#define overflowcalc(n, m, o) { /* n = result, m = accumulator, o = memory */ \
+    if (((n) ^ (uint16_t)(m)) & ((n) ^ (o)) & 0x0080) setoverflow();\
+        else clearoverflow();\
+}
+
+//a few general functions used by various other functions
+void push16(uint16_t pushval) {
+    write6502(BASE_STACK + sp, (pushval >> 8) & 0xFF);
+    write6502(BASE_STACK + ((sp - 1) & 0xFF), pushval & 0xFF);
+    sp -= 2;
+}
+
+void push8(uint8_t pushval) {
+    write6502(BASE_STACK + sp--, pushval);
+}
+
+uint16_t pull16() {
+    uint16_t temp16;
+    temp16 = read6502(BASE_STACK + ((sp + 1) & 0xFF)) | ((uint16_t)read6502(BASE_STACK + ((sp + 2) & 0xFF)) << 8);
+    sp += 2;
+    return(temp16);
+}
+
+uint8_t pull8() {
+    return (read6502(BASE_STACK + ++sp));
+}
+
+void reset6502() {
+    pc = (uint16_t)read6502(0xFFFC) | ((uint16_t)read6502(0xFFFD) << 8);
+    a = 0;
+    x = 0;
+    y = 0;
+    sp = 0xFD;
+    status |= FLAG_CONSTANT;
+}
+
