+++ cpu/tables.h
@@ -1,0 +1,61 @@
+/* Generated by buildtables.py */
+
+static void (*addrtable[256])() = {
+/*        |  0  |  1  |  2  |  3  |  4  |  5  |  6  |  7  |  8  |  9  |  A  |  B  |  C  |  D  |  E  |  F  |     */
+/* 0 */     imp, indx,  imp,  imp,   zp,   zp,   zp,  imp,  imp,  imm,  acc,  imp, abso, abso, abso,  imp, /* 0 */
+/* 1 */     rel, indy, ind0,  imp,   zp,  zpx,  zpx,  imp,  imp, absy,  acc,  imp, abso, absx, absx,  imp, /* 1 */
+/* 2 */    abso, indx,  imp,  imp,   zp,   zp,   zp,  imp,  imp,  imm,  acc,  imp, abso, abso, abso,  imp, /* 2 */
+/* 3 */     rel, indy, ind0,  imp,  zpx,  zpx,  zpx,  imp,  imp, absy,  acc,  imp, absx, absx, absx,  imp, /* 3 */
+/* 4 */     imp, indx,  imp,  imp,  imp,   zp,   zp,  imp,  imp,  imm,  acc,  imp, abso, abso, abso,  imp, /* 4 */
+/* 5 */     rel, indy, ind0,  imp,  imp,  zpx,  zpx,  imp,  imp, absy,  imp,  imp,  imp, absx, absx,  imp, /* 5 */
+/* 6 */     imp, indx,  imp,  imp,   zp,   zp,   zp,  imp,  imp,  imm,  acc,  imp,  ind, abso, abso,  imp, /* 6 */
+/* 7 */     rel, indy, ind0,  imp,  zpx,  zpx,  zpx,  imp,  imp, absy,  imp,  imp, ainx, absx, absx,  imp, /* 7 */
+/* 8 */     rel, indx,  imp,  imp,   zp,   zp,   zp,  imp,  imp,  imm,  imp,  imp, abso, abso, abso,  imp, /* 8 */
+/* 9 */     rel, indy, ind0,  imp,  zpx,  zpx,  zpy,  imp,  imp, absy,  imp,  imp, abso, absx, absx,  imp, /* 9 */
+/* A */     imm, indx,  imm,  imp,   zp,   zp,   zp,  imp,  imp,  imm,  imp,  imp, abso, abso, abso,  imp, /* A */
+/* B */     rel, indy, ind0,  imp,  zpx,  zpx,  zpy,  imp,  imp, absy,  imp,  imp, absx, absx, absy,  imp, /* B */
+/* C */     imm, indx,  imp,  imp,   zp,   zp,   zp,  imp,  imp,  imm,  imp,  imp, abso, abso, abso,  imp, /* C */
+/* D */     rel, indy, ind0,  imp,  imp,  zpx,  zpx,  imp,  imp, absy,  imp,  imp,  imp, absx, absx,  imp, /* D */
+/* E */     imm, indx,  imp,  imp,   zp,   zp,   zp,  imp,  imp,  imm,  imp,  imp, abso, abso, abso,  imp, /* E */
+/* F */     rel, indy, ind0,  imp,  imp,  zpx,  zpx,  imp,  imp, absy,  imp,  imp,  imp, absx, absx,  imp  /* F */
+};
+
+static void (*optable[256])() = {
+/*        |  0  |  1  |  2  |  3  |  4  |  5  |  6  |  7  |  8  |  9  |  A  |  B  |  C  |  D  |  E  |  F  |     */
+/* 0 */      brk,  ora,  nop,  nop,  tsb,  ora,  asl,  nop,  php,  ora,  asl,  nop,  tsb,  ora,  asl,  nop, /* 0 */
+/* 1 */      bpl,  ora,  ora,  nop,  trb,  ora,  asl,  nop,  clc,  ora,  inc,  nop,  trb,  ora,  asl,  nop, /* 1 */
+/* 2 */      jsr,  and,  nop,  nop,  bit,  and,  rol,  nop,  plp,  and,  rol,  nop,  bit,  and,  rol,  nop, /* 2 */
+/* 3 */      bmi,  and,  and,  nop,  bit,  and,  rol,  nop,  sec,  and,  dec,  nop,  bit,  and,  rol,  nop, /* 3 */
+/* 4 */      rti,  eor,  nop,  nop,  nop,  eor,  lsr,  nop,  pha,  eor,  lsr,  nop,  jmp,  eor,  lsr,  nop, /* 4 */
+/* 5 */      bvc,  eor,  eor,  nop,  nop,  eor,  lsr,  nop,  cli,  eor,  phy,  nop,  nop,  eor,  lsr,  nop, /* 5 */
+/* 6 */      rts,  adc,  nop,  nop,  stz,  adc,  ror,  nop,  pla,  adc,  ror,  nop,  jmp,  adc,  ror,  nop, /* 6 */
+/* 7 */      bvs,  adc,  adc,  nop,  stz,  adc,  ror,  nop,  sei,  adc,  ply,  nop,  jmp,  adc,  ror,  nop, /* 7 */
+/* 8 */      bra,  sta,  nop,  nop,  sty,  sta,  stx,  nop,  dey,  bit,  txa,  nop,  sty,  sta,  stx,  nop, /* 8 */
+/* 9 */      bcc,  sta,  sta,  nop,  sty,  sta,  stx,  nop,  tya,  sta,  txs,  nop,  stz,  sta,  stz,  nop, /* 9 */
+/* A */      ldy,  lda,  ldx,  nop,  ldy,  lda,  ldx,  nop,  tay,  lda,  tax,  nop,  ldy,  lda,  ldx,  nop, /* A */
+/* B */      bcs,  lda,  lda,  nop,  ldy,  lda,  ldx,  nop,  clv,  lda,  tsx,  nop,  ldy,  lda,  ldx,  nop, /* B */
+/* C */      cpy,  cmp,  nop,  nop,  cpy,  cmp,  dec,  nop,  iny,  cmp,  dex,  nop,  cpy,  cmp,  dec,  nop, /* C */
+/* D */      bne,  cmp,  cmp,  nop,  nop,  cmp,  dec,  nop,  cld,  cmp,  phx,  nop,  nop,  cmp,  dec,  nop, /* D */
+/* E */      cpx,  sbc,  nop,  nop,  cpx,  sbc,  inc,  nop,  inx,  sbc,  nop,  nop,  cpx,  sbc,  inc,  nop, /* E */
+/* F */      beq,  sbc,  sbc,  nop,  nop,  sbc,  inc,  nop,  sed,  sbc,  plx,  nop,  nop,  sbc,  inc,  nop  /* F */
+};
+
+static const uint32_t ticktable[256] = {
+/*        |  0  |  1  |  2  |  3  |  4  |  5  |  6  |  7  |  8  |  9  |  A  |  B  |  C  |  D  |  E  |  F  |     */
+/* 0 */       7,    6,    2,    2,    5,    3,    5,    2,    3,    2,    2,    2,    6,    4,    6,    2, /* 0 */
+/* 1 */       2,    5,    5,    2,    5,    4,    6,    2,    2,    4,    2,    2,    6,    4,    7,    2, /* 1 */
+/* 2 */       6,    6,    2,    2,    3,    3,    5,    2,    4,    2,    2,    2,    4,    4,    6,    2, /* 2 */
+/* 3 */       2,    5,    5,    2,    4,    4,    6,    2,    2,    4,    2,    2,    4,    4,    7,    2, /* 3 */
+/* 4 */       6,    6,    2,    2,    2,    3,    5,    2,    3,    2,    2,    2,    3,    4,    6,    2, /* 4 */
+/* 5 */       2,    5,    5,    2,    2,    4,    6,    2,    2,    4,    3,    2,    2,    4,    7,    2, /* 5 */
+/* 6 */       6,    6,    2,    2,    3,    3,    5,    2,    4,    2,    2,    2,    5,    4,    6,    2, /* 6 */
+/* 7 */       2,    5,    5,    2,    4,    4,    6,    2,    2,    4,    4,    2,    6,    4,    7,    2, /* 7 */
+/* 8 */       3,    6,    2,    2,    3,    3,    3,    2,    2,    2,    2,    2,    4,    4,    4,    2, /* 8 */
+/* 9 */       2,    6,    5,    2,    4,    4,    4,    2,    2,    5,    2,    2,    4,    5,    5,    2, /* 9 */
+/* A */       2,    6,    2,    2,    3,    3,    3,    2,    2,    2,    2,    2,    4,    4,    4,    2, /* A */
+/* B */       2,    5,    5,    2,    4,    4,    4,    2,    2,    4,    2,    2,    4,    4,    4,    2, /* B */
+/* C */       2,    6,    2,    2,    3,    3,    5,    2,    2,    2,    2,    2,    4,    4,    6,    2, /* C */
+/* D */       2,    5,    5,    2,    2,    4,    6,    2,    2,    4,    3,    2,    2,    4,    7,    2, /* D */
+/* E */       2,    6,    2,    2,    3,    3,    5,    2,    2,    2,    2,    2,    4,    4,    6,    2, /* E */
+/* F */       2,    5,    5,    2,    2,    4,    6,    2,    2,    4,    4,    2,    2,    4,    7,    2  /* F */
+};
